[fix/#32] toFloat 제거

+++ app/src/main/java/com/bongtu/baekseo/presentation/record/RecordDetailScreen.kt
@@ -167,7 +167,7 @@ private fun RecordDetailScreen(
                     .fillMaxWidth()
                     .clip(shape = RoundedCornerShape(size = 10.dp))
                     .background(color = BongBaekTheme.colors.gray750)
-                    .aspectRatio(MEMO_RATIO.toFloat()),
+                    .aspectRatio(MEMO_RATIO),
                 contentAlignment = Alignment.TopStart,
             ) {
                 Text(
